chore: Fix static analysis warnings
This change fixes most of the warnings/errors generated by the static
analysis tools used by this project (PMD, FindBugs and Checkstyle).

One FindBugs error that is not fixed by this commit is the usage of
`System.exit(int)` since we do not want to make the exiting logic to
complex for now.