feat(#144): Create marble announcer tile
This change adds an announcer tile which announces the next marble that
is going to be spawned.

I have created an AnnouncerActor which maps the identifier of the marble type
to an announcer sprite of the corresponding colour as tile sprite. I
have also added an announcer sprite for the joker ball.