feat(#146): Add teleporter tile
This branch will resolve issue #146 . It will add the teleporter feature where
a marble is able to teleport from one location of a teleporter track to the
other location.

EasyLevel has been modified to demonstrate the teleporter feature and has to be
changed back if the request is approved, before merging.

Closes #146