feat(#63): Add UX for pausing/resuming the game
This change adds a user interface for pausing and resuming the game.
A game is paused by pressing the escape button, which results in a pause
screen being showed to the user. This allows the user to click the
resume button or press the escape button again to resume the game

Improvements for the future include:
- Better design of the pause menu screen
- Allow the user to exit the game from the pause screen

Closes #63

Signed-off-by: Fabian Mastenbroek <fabianishere@outlook.com>